requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
loguru
requests
docker
APScheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
msal
slack_sdk
